pull out notificationContent from notifications list
