+++ assets/css/_notification_content.scss
@@ -1,0 +1,23 @@
+.m-notification-content__title-row {
+  display: flex;
+  justify-content: space-between;
+}
+
+.m-notification-content__title {
+  @include font-emph;
+  font-weight: 700;
+}
+
+.m-notification-content__age {
+  color: $color-font-grey;
+  @include font-small;
+}
+
+.m-notification-content__description {
+  margin-bottom: 0.5rem;
+  margin-top: 0.25rem;
+}
+
+.m-notification-content .m-properties-list {
+  padding: 0;
+}
