+++ assets/css/app.scss
@@ -24,6 +24,7 @@ $picker-width-wide: $picker-width-narrow + 3.1;
 @import "modal";
 @import "notifications";
 @import "notification_bell_icon";
+@import "notification_content";
 @import "notification_drawer";
 @import "page";
 @import "picker_container";
